-Added changes based on feedback -New damage potion + method to boost the player's damage -Super potion to boost all stats

+++ Assets/Scripts/Attack.cs
@@ -3,14 +3,14 @@ using System.Collections.Generic;
 using UnityEngine;
 /*
  * Written By: Gianni Coladonato
- * ID: 2414537
- * Last Modified (Date|ID): 09-11-2024, 2414537
+ * Date Created: 09-11-2024 | Last Modified: 13-04-2025
+ * 
+ * This script is used to allow entities to attack each other
  */
-//This script allows for entities to attack and deal damage to one another
 public class Attack : MonoBehaviour
 {
     //The damage an attack deals, can be adjusted in the inspector
-    public int damage = 10;
+    public float damage = 10;
     //The knockback an attack applies on hit
     public Vector2 knockback = Vector2.zero;
 
@@ -23,7 +23,7 @@ public class Attack : MonoBehaviour
             //Flip the knockback direction if the player is facing another direction
             Vector2 directionalKnock = transform.parent.localScale.x > 0 ? knockback : new Vector2(-knockback.x, knockback.y);
             //Hit the target
-            damageable.Hit(damage, directionalKnock);
+            damageable.Hit((int)damage, directionalKnock);
             //Debug.Log($"{collision.name} hit for {damage}");
         }
     }
